format source labels at data load (eventually should be done in the data repos already)
