+++ dplace_app/loader/phylogenies.py
@@ -40,17 +40,14 @@ def _load(obj, sources, sequences):
     Tree(tree.newick_string, format=1)
 
     for item in obj.taxa:
-        name_on_tip = item['taxon']
-        xd_ids = nfilter(i.strip() for i in item['xd_ids'].split(','))
-        society_ids = nfilter(i.strip() for i in item['soc_ids'].split(','))
-        if not xd_ids:
+        if not item.xd_ids:
             continue
 
-        label = LanguageTreeLabels.objects.create(languageTree=tree, label=name_on_tip)
+        label = LanguageTreeLabels.objects.create(languageTree=tree, label=item.taxon)
         tree.taxa.add(label)
-        for society in Society.objects.all().filter(xd_id__in=xd_ids):
+        for society in Society.objects.all().filter(xd_id__in=item.xd_ids):
             try:
-                f_order = len(society_ids) - society_ids.index(society.ext_id) - 1
+                f_order = len(item.soc_ids) - item.soc_ids.index(society.ext_id) - 1
             except:
                 f_order = 0
             sequences.append(LanguageTreeLabelsSequence(
