+++ dplace_app/tests/data/datasets/dscultural/data.csv
@@ -1,6 +1,6 @@
 soc_id,sub_case,year,var_id,code,comment,references,source_coded_data,admin_comment
-society1,Gei/Khauan tribe,1840,1,1,,"Schultze, 1907; Name, 1924",EthnographicAtlas_1967_p62,
-society2,,1920,1,4,,"Name, 1894",Ethnology_Vol1_No3_Jul_1962,
+society1,Gei/Khauan tribe,1840,1,1,,"schultze1907; name1924",EthnographicAtlas_1967_p62,
+society2,,1920,1,4,,"name1894",Ethnology_Vol1_No3_Jul_1962,
 society3,"Epulu net-hunters, Ituri forest",1930,1,3,,"",EthnographicAtlas_1967_p62,
-society1,Gei/Khauan tribe,1840,2,56.8,,"Schultze, 1907; Name, 1924",EthnographicAtlas_1967_p62,
-society2,,1920,2,250.4,,"Name, 1894",Ethnology_Vol1_No3_Jul_1962,
+society1,Gei/Khauan tribe,1840,2,56.8,,"schultze1907; name1924",EthnographicAtlas_1967_p62,
+society2,,1920,2,250.4,,"name1894",Ethnology_Vol1_No3_Jul_1962,
