--- dplace_app/tests/data/datasets/dscultural/references.csv
@@ -1,5 +1,0 @@
-key,citation
-"Name, 1924","Name, M. 1924. Title. Place."
-"one","two","too many cols"
-"Name no comma 1924","Name, M. 1924. Title. Place."
-"Name, 1894","Name, N. L. von. 1894. Title. More. Info. 22: 43-44. ."
