+++ dplace_app/tests/data/datasets/sources.bib
@@ -1,0 +1,12 @@
+@misc{name1894,
+    author = {Name},
+    year = {1894},
+    title = {Title. More. Info. 22: 43-44.},
+    key = {Name, 1894}
+}
+@misc{name1924,
+    author = {Name, N. L. von},
+    year = {1924},
+    title = {Title},
+    key = {Name, 1924}
+}
